[API] Move OrderItemsTest to disabled test directory
